+++ FBSDKCoreKit/FBSDKCoreKit/include/FBSDKError.h
@@ -11,6 +11,7 @@
 NS_ASSUME_NONNULL_BEGIN
 
 NS_SWIFT_NAME(SDKError)
+DEPRECATED_MSG_ATTRIBUTE("`SDKError` is deprecated and will be removed in the next major release; use `ErrorFactory` and/or `NetworkErrorChecker` instead")
 @interface FBSDKError : NSObject
 
 + (NSError *)errorWithCode:(NSInteger)code message:(nullable NSString *)message;